Resolve import cycle by moving tests to separate module
