+++ store/types/iterator_test.go
@@ -1,22 +1,24 @@
-package pageiterator
+package types_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	dbm "github.com/tendermint/tm-db"
 
-	"github.com/cosmos/cosmos-sdk/store/transient"
+	"github.com/cosmos/cosmos-sdk/store/iavl"
 	"github.com/cosmos/cosmos-sdk/store/types"
 )
 
-func newMemTestKVStore() types.KVStore {
-	// rootmulti.NewStore/iavl.LoadStore introduces circular dependency if alias for KVStorePrefixIteratorPaginated is added to types/ module.
-	// both modules import store/errors and store/errors is an alias for some of the errors in types/errors.go
-	return transient.NewStore()
+func newMemTestKVStore(t *testing.T) types.KVStore {
+	db := dbm.NewMemDB()
+	store, err := iavl.LoadStore(db, types.CommitID{}, types.PruneNothing, false)
+	require.NoError(t, err)
+	return store
 }
 
 func TestPaginatedIterator(t *testing.T) {
-	kvs := newMemTestKVStore()
+	kvs := newMemTestKVStore(t)
 	total := 10
 	lth := total - 1
 	asc := make([][]byte, total)
@@ -83,15 +85,17 @@ func TestPaginatedIterator(t *testing.T) {
 		t.Run(tc.desc, func(t *testing.T) {
 			var iter types.Iterator
 			if tc.reverse {
-				iter = KVStoreReversePrefixIteratorPaginated(kvs, nil, tc.page, tc.limit)
+				iter = types.KVStoreReversePrefixIteratorPaginated(kvs, nil, tc.page, tc.limit)
 			} else {
-				iter = KVStorePrefixIteratorPaginated(kvs, nil, tc.page, tc.limit)
+				iter = types.KVStorePrefixIteratorPaginated(kvs, nil, tc.page, tc.limit)
 			}
 			defer iter.Close()
+
 			result := [][]byte{}
 			for ; iter.Valid(); iter.Next() {
 				result = append(result, iter.Key())
 			}
+
 			require.Equal(t, tc.result, result)
 			require.False(t, iter.Valid())
 		})
@@ -99,13 +103,16 @@ func TestPaginatedIterator(t *testing.T) {
 }
 
 func TestPaginatedIteratorPanicIfInvalid(t *testing.T) {
-	kvs := newMemTestKVStore()
-	iter := KVStorePrefixIteratorPaginated(kvs, nil, 1, 1)
+	kvs := newMemTestKVStore(t)
+
+	iter := types.KVStorePrefixIteratorPaginated(kvs, nil, 1, 1)
 	defer iter.Close()
 	require.False(t, iter.Valid())
 	require.Panics(t, func() { iter.Next() }) // "iterator is empty"
+
 	kvs.Set([]byte{1}, []byte{})
-	iter = KVStorePrefixIteratorPaginated(kvs, nil, 1, 0)
+
+	iter = types.KVStorePrefixIteratorPaginated(kvs, nil, 1, 0)
 	defer iter.Close()
 	require.False(t, iter.Valid())
 	require.Panics(t, func() { iter.Next() }) // "not empty but limit is zero"
